fix(web): make playground compile under webdev
