+++ pkgs/liquify/lib/src/fs/fs_stub.dart
@@ -1,0 +1,1 @@
+// Stub export for platforms without io or js_interop support.
