+++ pkgs/liquify/pubspec.yaml
@@ -13,7 +13,6 @@ topics:
   - teamplate-engine
   - dart
 
-resolution: workspace
 environment:
   sdk: ">=3.8.0 <4.0.0"
 
@@ -28,6 +27,7 @@ dependencies:
   file: ^7.0.1
   args: ^2.7.0
   logging: ^1.3.0
+  web: ^1.1.1
 
 dev_dependencies:
   lints: ^6.1.0
@@ -35,6 +35,8 @@ dev_dependencies:
   stack_trace: ^1.12.1
   test: ^1.31.1
   coverage: ^1.15.0
+  build_runner: ^2.15.0
+  build_web_compilers: ^4.8.0
 
 funding:
   - https://www.buymeacoffee.com/kingwill101
